the great opaque pointer migration
LLVM moved to opaque pointers a good while ago, and it turns out I've
been using an outdated version of the LLVM C api.

This commit makes all exprs keep track of their types so that type
information can be supplied to the llvm api with a simple `ast_type` to
`LLVMTypeRef` conversion.

This commit also introduces the idea of 'ownership' to types and subtypes for
ast_exprs and ast_types respectively. If you don't own a type, you don't free that
type. It belongs to someone else!
	This helps cut out some unnecessary ast_type allocations.
	Sharing is caring.